1dollar-challenge/requirements.txt
python-docx==0.8.11
fpdf==1.7.2
streamlit==1.32.0 